requirements.dev.txt
flake8>=3.9.2,<3.10
django-debug-toolbar==3.7.0